+++ concrete/db/orm/__init__.py
@@ -1,0 +1,6 @@
+from . import models, schemas
+from .setup import SessionLocal, engine
+
+models.Base.metadata.create_all(bind=engine)
+
+__all__ = ["models", "schemas", "SessionLocal"]
